Refactored shared constants and functions into a seperate file
