+++ src/scripts/candidate_calculation_industry.py
@@ -6,8 +6,7 @@ import pathlib
 
 import pandas as pd
 
-CONTRIBUTION_TYPE = "RCPT"
-TYPE_COLUMN = "Rec_Type"
+from shared_calculations import CONTRIBUTION_TYPE, TYPE_COLUMN
 
 
 def raw_industry_contributions(spreadsheet_path):
